pyproject.toml
[project]
name = "doc-management-system"
version = "0.1.0"
description = "FastAPI-based REST API for managing products"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.6",
    "uvicorn[standard]==0.32.1",
    "motor==3.6.0",
    "pydantic==2.10.3",
    "python-multipart==0.0.17",
    "jinja2==3.1.4",
    "python-dotenv==1.0.1",
    "firebase-admin>=7.3.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
